- Add carrot UI

+++ BerlinMiniGameJam-Dec/Assets/Scripts/CarrotCollector.cs
@@ -5,7 +5,7 @@ public class CarrotCollector : MonoBehaviour
 {
 	[SerializeField]
 	float carrotTimer = 4.0f;
-	float requiredCarrotsForAttacking = 4.0f;
+	int requiredCarrotsForAttacking = 4;
 	
 	private int carrotCounter = 0;
